continuing the porting of the math utilities, almost at the end
