+++ src/data_structures/Matrix.cxx
@@ -1,5 +1,6 @@
 #include "data_structures/Matrix.hxx"
 #include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 #include <iostream>
 #include <stdexcept>
 #include <vector>
