+++ src/data_structures/Matrix.hxx
@@ -142,11 +142,6 @@ class Matrix {
         void printMatrix()const;
 
 
-        bool approximatelyEqual(float a, float b, float epsilon)
-        {
-            return fabs(a - b) <= ( (fabs(a) < fabs(b) ? fabs(b) : fabs(a)) * epsilon);
-        }
-
     protected:
         int rows_, cols_;
         T *_matrix=nullptr;
