+++ src/data_structures/WeightedEdgeGraph.cxx
@@ -1,7 +1,7 @@
 #include "data_structures/WeightedEdgeGraph.hxx"
 #include "data_structures/Matrix.hxx"
 #include "utils/utilities.hxx"
-//#include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 #include <algorithm>
 #include <iterator>
 #include <map>
