+++ src/mains/ComputationTesting.cxx
@@ -1,5 +1,6 @@
 #include "data_structures/WeightedEdgeGraph.hxx"
 #include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 #include "computation/Computation.hxx"
 #include <iostream>
