+++ src/mains/main.cxx
@@ -19,6 +19,7 @@
 #include "computation/DissipationModelScaled.hxx"
 #include "data_structures/WeightedEdgeGraph.hxx"
 #include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 #include "CustomFunctions.hxx"
 #include "logging/Logger.hxx"
