+++ src/testing/ComputationTesting.cc
@@ -5,6 +5,7 @@
 #include "computation/Computation.hxx"
 #include "data_structures/Matrix.hxx"
 #include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 #include "computation/DissipationModel.hxx"
 #include "computation/DissipationModelScaled.hxx"
 #include "computation/ConservationModel.hxx"
