+++ src/testing/ComputationTestingPerturbation.cc
@@ -12,6 +12,7 @@
 #include "data_structures/WeightedEdgeGraph.hxx"
 #include "data_structures/Matrix.hxx"
 #include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 
 class ComputationTestingPerturbation : public ::testing::Test {
     protected:
