+++ src/testing/utilitiesTesting.cc
@@ -3,6 +3,7 @@
 #include <string>
 #include <vector>
 #include "utils/utilities.hxx"
+#include "utils/mathUtilities.hxx"
 
 class utilitiesTesting : public ::testing::Test {
     protected:
